Moved SmartEnum to a separate folder and moved OrderedSmartEnum to it's own file.
