+++ CyberCavsScoutingSystem/WPFUtilities/SmartEnum/SmartEnum.cs
@@ -86,67 +86,3 @@ public abstract class SmartEnum<T> : IEquatable<SmartEnum<T>> where T : SmartEnu
 	}
 
 }
-
-public abstract class OrderedSmartEnum<T> : SmartEnum<T>, IComparable where T : OrderedSmartEnum<T> {
-
-	protected OrderedSmartEnum(string name, int value) : base(name, value) { }
-
-	public new static IEnumerable<T> GetOptions() {
-
-		return SmartEnum<T>.GetOptions().OrderBy(x => x.Value);
-
-		//Verbose version
-		//IEnumerable<T> allOptions = SmartEnum<T>.GetAllOptions();
-		//IEnumerable<T> sortedOptions = allOptions.OrderBy(x => x.Value);
-		//return sortedOptions;
-	}
-
-	public int CompareTo(object? obj) {
-
-		if (obj is OrderedSmartEnum<T> other) {
-			return CompareTo(other);
-		}
-
-		throw new ArgumentException($"The parameter {nameof(obj)} could not be compared this object of type {typeof(OrderedSmartEnum<T>)}");
-	}
-
-	public int CompareTo(OrderedSmartEnum<T>? other) {
-
-		ArgumentNullException.ThrowIfNull(other, nameof(other));
-
-		return Value.CompareTo(other.Value);
-	}
-
-	public static bool operator <(OrderedSmartEnum<T>? left, OrderedSmartEnum<T>? right) {
-
-		ArgumentNullException.ThrowIfNull(left, nameof(left));
-		ArgumentNullException.ThrowIfNull(right, nameof(right));
-
-		return left.CompareTo(right) < 0;
-	}
-
-	public static bool operator >(OrderedSmartEnum<T>? left, OrderedSmartEnum<T>? right) {
-
-		ArgumentNullException.ThrowIfNull(left, nameof(left));
-		ArgumentNullException.ThrowIfNull(right, nameof(right));
-
-		return left.CompareTo(right) > 0;
-	}
-
-	public static bool operator <=(OrderedSmartEnum<T>? left, OrderedSmartEnum<T>? right) {
-
-		ArgumentNullException.ThrowIfNull(left, nameof(left));
-		ArgumentNullException.ThrowIfNull(right, nameof(right));
-
-		return left.CompareTo(right) <= 0;
-	}
-
-	public static bool operator >=(OrderedSmartEnum<T> left, OrderedSmartEnum<T>? right) {
-
-		ArgumentNullException.ThrowIfNull(left, nameof(left));
-		ArgumentNullException.ThrowIfNull(right, nameof(right));
-
-		return left.CompareTo(right) >= 0;
-	}
-
-}
